+++ src/distance.cpp
@@ -1,8 +1,8 @@
 #include "distance.h"
 
-double euclidean_distance(struct point x1, struct point x2)
+float euclidean_distance(struct point x1, struct point x2)
 {
-    double dist = 0;
+    float dist = 0;
     for (unsigned int i=0; i < x1.dim; i++)
         dist += pow((x1.cord[i] - x2.cord[i]), 2);
 
@@ -10,9 +10,9 @@ double euclidean_distance(struct point x1, struct point x2)
 }
 
 
-double manhattan_distance(struct point x1, struct point x2)
+float manhattan_distance(struct point x1, struct point x2)
 {
-    double dist = 0;
+    float dist = 0;
     for (unsigned int i=0; i < x1.dim; i++)  
         dist += abs(x1.cord[i] - x2.cord[i]);
